requirement.txt
aenum==3.1.15
aioboto3==12.1.0
aiobotocore==2.8.0
aiohttp==3.9.0
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
aws-wsgi==0.2.7
black==23.12.0
blinker==1.7.0
boto3==1.33.1
boto3-type-annotations==0.3.1
botocore==1.33.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Deprecated==1.2.14
exceptiongroup==1.2.0
Flask==3.0.0
fonttools==4.47.0
frozenlist==1.4.0
future==0.18.3
google-auth==2.25.2
google-auth-oauthlib==1.1.0
gspread==5.10.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.2
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.5
line-bot-sdk==2.3.0
loguru==0.7.2
MarkupSafe==2.1.3
matplotlib==3.8.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
pip-check==2.8.1
platformdirs==4.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.8.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
terminaltables==3.1.10
tomli==2.0.1
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==1.26.18
uvloop==0.19.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
